pyproject.toml
[tool.poetry]
name = "general-llm"
version = "0.1.0"
description = ""
authors = ["takaharaomi <osmith8310@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.115.5"
uvicorn = {extras = ["standard"], version = "^0.32.0"}
python-dotenv = "^1.0.1"
langchain-core = "^0.3.19"
langchain-openai = "^0.2.9"
tavily-python = "^0.5.0"
langchain-community = "^0.3.8"
black = "^24.10.0"


[tool.poetry.group.dev.dependencies]
mypy = "^1.13.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
